% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data_from_url.R
\name{download_data_from_url}
\alias{download_data_from_url}
\title{download_data_from_url}
\usage{
download_data_from_url(
  download_url,
  format = "JSON",
  try_unsupported_format = FALSE
)
}
\arguments{
\item{download_url}{string. a valid url}

\item{format}{string.}

\item{try_unsupported_format}{logical. default FALSE.}
}
\description{
download_data_from_url
}
