% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_package_metadata.R
\name{get_package_metadata}
\alias{get_package_metadata}
\title{get_package_metadata}
\usage{
get_package_metadata("ladestationen-fuer-elektroautos")
}
\arguments{
\item{package_name}{a string.}
}
\description{
Get package metadata
}
